Move commonly used filters into server module and create when setting up jersey. Add CORS filter but still pending confirmation on matching logic.
