--- jaxrs-service/src/main/java/com/quorum/tessera/api/filter/DomainFilter.java
@@ -1,14 +1,0 @@
-package com.quorum.tessera.api.filter;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import javax.ws.rs.NameBinding;
-
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(value = RetentionPolicy.RUNTIME)
-@NameBinding
-public @interface DomainFilter {
-
-}
