+++ jaxrs-service/src/main/java/com/quorum/tessera/thridparty/RawTransactionResource.java
@@ -1,6 +1,5 @@
 package com.quorum.tessera.thridparty;
 
-import com.quorum.tessera.api.filter.DomainFilter;
 import com.quorum.tessera.api.model.*;
 import com.quorum.tessera.config.apps.ThirdPartyApp;
 import com.quorum.tessera.transaction.TransactionManager;
@@ -23,7 +22,6 @@ import static javax.ws.rs.core.MediaType.*;
 /**
  * Provides endpoints for dealing with raw transactions
  */
-@DomainFilter
 @Path("/")
 public class RawTransactionResource implements ThirdPartyApp {
